remove s/c verification by external code size
